Some more unittests, small restructuring for tests with mocked python modules
